Authentication

+++ src/actions/actions.ts
@@ -1,13 +1,29 @@
 "use server";
 
 import prisma from "@/lib/db";
+import { getKindeServerSession } from "@kinde-oss/kinde-auth-nextjs/server";
 import { revalidatePath } from "next/cache";
+import { redirect } from "next/navigation";
 
 export async function addExpense(formData: FormData) {
+    // Auth check
+    const { isAuthenticated, getUser } = getKindeServerSession();
+
+    if (!(await isAuthenticated())) {
+        return redirect('/api/auth/login');
+    }
+
+    const user = await getUser();
+
+    if (!user) {
+        throw new Error("User not found");
+    }
+
     await prisma.expense.create({
         data: {
             description: formData.get("description") as string,
             amount: parseFloat(formData.get("amount") as string),
+            creatorId: user.id,
         },
     });
 
@@ -15,6 +31,13 @@ export async function addExpense(formData: FormData) {
 }
 
 export async function deleteExpense(id: number) {
+    // Auth check
+    const { isAuthenticated } = getKindeServerSession();
+
+    if (!(await isAuthenticated())) {
+        return redirect('/api/auth/login');
+    }
+
     await prisma.expense.delete({
         where: {
             id,
